Add SerializerAsync
