+++ Packages/UGF.Serialize/Runtime/SerializerAsync.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: 60cbb402061e4b54adf2c030bd3b8c08
-timeCreated: 1602519818
+guid: 5402f1e00a8040208f1f8e37c75cd364
+timeCreated: 1642774013
