+++ Packages/UGF.Serialize/Runtime/SerializerAsync`1.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 60cbb402061e4b54adf2c030bd3b8c08
+timeCreated: 1602519818
